Separate what a name is from how a name is written
A quantity names its units in these elements, so quantity.ts reaches for the
type and for atom, and the writing of an element can never reach back for the
quantity it would have to write. That is what a caption needs to be in the
reader's own units rather than in whichever ones the script was read in.

So the type and its one constructor sit on their own, where the quantity module
and the writing of a name can both reach them, and nothing has to import a name
to say what one is. Re-exporting them from where they were would have spared
the churn, but this codebase puts export on the declaration.
