+++ react/cf-og-worker/src/data.ts
@@ -25,7 +25,8 @@ import { Table } from '../../src/urban-stats-script/constants/table'
 import { deriveConditionLabel, deriveMapLabel } from '../../src/urban-stats-script/derive-human-readable-name'
 import { createRequestExecutor } from '../../src/urban-stats-script/execute-request'
 import { geometry } from '../../src/utils/geometry'
-import { HumanReadableName, reifyString } from '../../src/utils/human-readable-name'
+import { HumanReadableName } from '../../src/utils/human-readable-element'
+import { reifyString } from '../../src/utils/human-readable-name'
 import { Feature } from '../../src/utils/protos'
 import { StoredUnit } from '../../src/utils/quantity'
 import { loadFeatureFromPossibleSymlink } from '../../src/utils/symlinks'
