+++ react/cf-og-worker/src/embed.tsx
@@ -15,7 +15,8 @@ import { pieSlicePath, pieSlices } from '../../src/syau/cluster-geometry'
 import { Inset } from '../../src/urban-stats-script/constants/insets'
 import { mixWithBackground } from '../../src/utils/color'
 import { computeAspectRatioForInsets } from '../../src/utils/coordinates'
-import { HumanReadableName, reifyString } from '../../src/utils/human-readable-name'
+import { HumanReadableName } from '../../src/utils/human-readable-element'
+import { reifyString } from '../../src/utils/human-readable-name'
 import { StoredUnit } from '../../src/utils/quantity'
 import { classifyStatistic, unitTypeToStoredUnit } from '../../src/utils/unit'
 import logoSvg from '../assets/logo.svg'
