+++ react/src/components/csv-export.tsx
@@ -7,7 +7,8 @@ import React, { ReactNode, useCallback, useContext } from 'react'
 import { Settings } from '../page_template/settings'
 import { groupYearKeys, StatGroupSettings } from '../page_template/statistic-settings'
 import { USSOpaqueValue, USSValue } from '../urban-stats-script/types-values'
-import { HumanReadableName, reifyString } from '../utils/human-readable-name'
+import { HumanReadableName } from '../utils/human-readable-element'
+import { reifyString } from '../utils/human-readable-name'
 import { sanitize } from '../utils/paths'
 import { Article } from '../utils/protos'
