+++ react/src/components/table.tsx
@@ -12,7 +12,8 @@ import { MobileArticlePointers, rowExpandedKey, useSetting, useSettings } from '
 import { Universe, useDefinedUniverse, useUniverse } from '../universe'
 import { withButtonRole } from '../utils/a11y'
 import { assert } from '../utils/defensive'
-import { HumanReadableName, reifyReact, reifyString } from '../utils/human-readable-name'
+import { HumanReadableName } from '../utils/human-readable-element'
+import { reifyReact, reifyString } from '../utils/human-readable-name'
 import { sanitize } from '../utils/paths'
 import { useComparisonHeadStyle, useMobileLayout } from '../utils/responsive'
 import { isAllowedToBeShown } from '../utils/restricted-types'
