+++ react/src/mapper/components/Colorbar.tsx
@@ -5,7 +5,8 @@ import { renderInequality } from '../../components/unit-display'
 import { Colors } from '../../page_template/color-themes'
 import { useColors } from '../../page_template/colors'
 import { ScaleInstance } from '../../urban-stats-script/constants/scale'
-import { HumanReadableName, reifyReact } from '../../utils/human-readable-name'
+import { HumanReadableName } from '../../utils/human-readable-element'
+import { reifyReact } from '../../utils/human-readable-name'
 import { StoredUnit } from '../../utils/quantity'
 import { rampColorer } from '../map-rendering'
 import { Keypoints } from '../ramps'
