+++ react/src/mapper/map-generator.tsx
@@ -31,7 +31,7 @@ import { loadImage } from '../utils/Image'
 import { editIndex, EditSeq } from '../utils/array-edits'
 import { computeAspectRatioForInsets } from '../utils/coordinates'
 import { makeDebugLogger } from '../utils/debug-logging'
-import { HumanReadableName } from '../utils/human-readable-name'
+import { HumanReadableName } from '../utils/human-readable-element'
 import { ICoordinate } from '../utils/protos'
 import { StoredUnit } from '../utils/quantity'
 import { plainNumber, unitTypeToStoredUnit } from '../utils/unit'
