+++ react/src/mapper/settings/Selector.tsx
@@ -14,7 +14,8 @@ import { emptyLocation } from '../../urban-stats-script/lexer'
 import { constantString, parseNoErrorAsCustomNode, parseNoErrorAsExpression } from '../../urban-stats-script/parser'
 import { Documentation, TypeEnvironment, USSType } from '../../urban-stats-script/types-values'
 import { TestUtils } from '../../utils/TestUtils'
-import { HumanReadableName, reifyReact, reifyString } from '../../utils/human-readable-name'
+import { HumanReadableName } from '../../utils/human-readable-element'
+import { reifyReact, reifyString } from '../../utils/human-readable-name'
 import { useTextAreaSizeSync } from '../../utils/text-area-size-sync'
 
 import * as l from './../../urban-stats-script/literal-parser'
