+++ react/src/stat/AddColumnSearchBox.tsx
@@ -5,7 +5,8 @@ import { possibilities } from '../mapper/settings/parseExpr'
 import { Selection } from '../mapper/settings/selector-classifier'
 import { addColumn } from '../urban-stats-script/add-column'
 import { TypeEnvironment } from '../urban-stats-script/types-values'
-import { HumanReadableName, reifyReact, reifyString } from '../utils/human-readable-name'
+import { HumanReadableName } from '../utils/human-readable-element'
+import { reifyReact, reifyString } from '../utils/human-readable-name'
 
 import { Statistic, StatSetter } from './types'
 import { mapUSSFromStat } from './utils'
