+++ react/src/utils/human-readable-name.tsx
@@ -1,14 +1,7 @@
 import React, { ReactNode } from 'react'
 
 import { codeStyle } from './code-style'
-
-export type HumanReadableElement = { type: 'atom', value: string } | { type: 'code', value: string } | { type: 'where' | 'superscript' | 'subscript' | 'parens', value: HumanReadableElement[] }
-
-export type HumanReadableName = string | HumanReadableElement[]
-
-export function atom(value: string): HumanReadableElement[] {
-    return [{ type: 'atom', value }]
-}
+import { HumanReadableElement, HumanReadableName } from './human-readable-element'
 
 export function reifyReact(elements: HumanReadableElement[] | string): ReactNode {
     if (typeof elements === 'string') return elements
